feat(status): Checking status of annotated applications only (#293)
